chore: DRY up the two codegen subclasses (#99)
* chore: DRY up the two codegens

* chore: pipe through custom headers and footers
