+++ src/code-generator/PlaywrightCodeGenerator.js
@@ -1,4 +1,3 @@
-import domEvents from './dom-events-to-record'
 import pptrActions from './pptr-actions'
 import Block from './Block'
 import CodeGenerator from './CodeGenerator'
@@ -18,146 +17,19 @@ const wrappedFooter = `  ${footer}
 })()`
 
 export default class PlaywrightCodeGenerator extends CodeGenerator {
+  constructor (options) {
+    super(options)
+    this._header = header
+    this._wrappedHeader = wrappedHeader
+    this._footer = footer
+    this._wrappedFooter = wrappedFooter
+  }
+
   generate (events) {
     return importPlaywright + this._getHeader() + this._parseEvents(events) + this._getFooter()
   }
 
-  _getHeader () {
-    console.debug(this._options)
-    let hdr = this._options.wrapAsync ? wrappedHeader : header
-    hdr = this._options.headless ? hdr : hdr.replace('launch()', 'launch({ headless: false })')
-    return hdr
-  }
-
-  _getFooter () {
-    return this._options.wrapAsync ? wrappedFooter : footer
-  }
-
-  _parseEvents (events) {
-    console.debug(`generating code for ${events ? events.length : 0} events`)
-    let result = ''
-
-    if (!events) return result
-
-    for (let i = 0; i < events.length; i++) {
-      const { action, selector, value, href, keyCode, tagName, frameId, frameUrl } = events[i]
-
-      // we need to keep a handle on what frames events originate from
-      this._setFrames(frameId, frameUrl)
-
-      switch (action) {
-        case 'keydown':
-          if (keyCode === 9) { // tab key
-            this._blocks.push(this._handleKeyDown(selector, value, keyCode))
-          }
-          break
-        case 'click':
-          this._blocks.push(this._handleClick(selector, events))
-          break
-        case 'change':
-          if (tagName === 'SELECT') {
-            this._blocks.push(this._handleChange(selector, value))
-          }
-          break
-        case pptrActions.GOTO:
-          this._blocks.push(this._handleGoto(href, frameId))
-          break
-        case pptrActions.VIEWPORT:
-          this._blocks.push((this._handleViewport(value.width, value.height)))
-          break
-        case pptrActions.NAVIGATION:
-          this._blocks.push(this._handleWaitForNavigation())
-          this._hasNavigation = true
-          break
-        case pptrActions.SCREENSHOT:
-          this._blocks.push(this._handleScreenshot(value))
-          break
-      }
-    }
-
-    if (this._hasNavigation && this._options.waitForNavigation) {
-      console.debug('Adding navigationPromise declaration')
-      const block = new Block(this._frameId, { type: pptrActions.NAVIGATION_PROMISE, value: 'const navigationPromise = page.waitForNavigation()' })
-      this._blocks.unshift(block)
-    }
-
-    console.debug('post processing blocks:', this._blocks)
-    this._postProcess()
-
-    const indent = this._options.wrapAsync ? '  ' : ''
-    const newLine = `\n`
-
-    for (let block of this._blocks) {
-      const lines = block.getLines()
-      for (let line of lines) {
-        result += indent + line.value + newLine
-      }
-    }
-
-    return result
-  }
-
-  _handleKeyDown (selector, value) {
-    const block = new Block(this._frameId)
-    block.addLine({ type: domEvents.KEYDOWN, value: `await ${this._frame}.type('${selector}', '${value}')` })
-    return block
-  }
-
-  _handleClick (selector) {
-    const block = new Block(this._frameId)
-    if (this._options.waitForSelectorOnClick) {
-      block.addLine({ type: domEvents.CLICK, value: `await ${this._frame}.waitForSelector('${selector}')` })
-    }
-    block.addLine({ type: domEvents.CLICK, value: `await ${this._frame}.click('${selector}')` })
-    return block
-  }
-
-  _handleChange (selector, value) {
-    return new Block(this._frameId, { type: domEvents.CHANGE, value: `await ${this._frame}.select('${selector}', '${value}')` })
-  }
-
-  _handleGoto (href) {
-    return new Block(this._frameId, { type: pptrActions.GOTO, value: `await ${this._frame}.goto('${href}')` })
-  }
-
   _handleViewport (width, height) {
     return new Block(this._frameId, { type: pptrActions.VIEWPORT, value: `await ${this._frame}.setViewportSize({ width: ${width}, height: ${height} })` })
-  }
-
-  _handleScreenshot (options) {
-    let block
-
-    if (options && options.x && options.y && options.width && options.height) {
-      // remove the tailing 'px'
-      for (let prop in options) {
-        if (options.hasOwnProperty(prop) && options[prop].slice(-2) === 'px') {
-          options[prop] = options[prop].substring(0, options[prop].length - 2)
-        }
-      }
-
-      block = new Block(this._frameId, {
-        type: pptrActions.SCREENSHOT,
-        value: `await ${this._frame}.screenshot({ path: 'screenshot_${this._screenshotCounter}.png', clip: { x: ${options.x}, y: ${options.y}, width: ${options.width}, height: ${options.height} } })` })
-    } else {
-      block = new Block(this._frameId, { type: pptrActions.SCREENSHOT, value: `await ${this._frame}.screenshot({ path: 'screenshot_${this._screenshotCounter}.png' })` })
-    }
-
-    this._screenshotCounter++
-    return block
-  }
-
-  _postProcessSetFrames () {
-    for (let [i, block] of this._blocks.entries()) {
-      const lines = block.getLines()
-      for (let line of lines) {
-        if (line.frameId && Object.keys(this._allFrames).includes(line.frameId.toString())) {
-          const declaration = `const frame_${line.frameId} = frames.find(f => f.url() === '${this._allFrames[line.frameId]}')`
-          this._blocks[i].addLineToTop(({ type: pptrActions.FRAME_SET, value: declaration }))
-          this._blocks[i].addLineToTop({ type: pptrActions.FRAME_SET, value: 'let frames = await page.frames()' })
-          delete this._allFrames[line.frameId]
-          break
-        }
-      }
-    }
   }
 }
